Moved RC class to vision_inference
