+++ VisionTrainingGround/RCnet/src/main.py
@@ -6,7 +6,7 @@ This script initializes and runs the training pipeline
 
 import argparse
 import torch
-from region_classifier import ImageClassifier
+from vision_inference.region_classifier import RegionClassifier
 
 
 def parse_args():
@@ -67,7 +67,7 @@ if __name__ == "__main__":
     torch.cuda.empty_cache()
 
     # Create the classifier object
-    classifier = ImageClassifier(
+    classifier = RegionClassifier(
         data_path=args.data_dir,
         save_plot_flag=args.save_plot_flag,
         save_plot_path=args.save_plot_path,
